Remove all backwards compatibility and establish single golden path
This commit completes the architecture cleanup by:

1. **REMOVED InfrastructureManager entirely**
   - Deleted buttermilk/_core/infrastructure.py
   - Removed all InfrastructureManager imports/exports from __init__.py
   - ExecutionContext now handles all infrastructure directly

2. **SIMPLIFIED ExecutionContext to use service-aware pattern**
   - Removed deprecated fields: secret_provider, pubsub
   - Added _find_cloud_with_service() method
   - Updated secret_manager property to use service-aware clouds[].secrets
   - Fixed all references to use the new pattern

3. **CLEANED configuration bootstrap**
   - Removed deprecated secret_provider and pubsub extraction
   - Added cloud configuration instantiation using hydra.utils.instantiate()
   - Updated method signatures to return only ExecutionContext
   - Fixed all callers to use simplified return values

4. **REMOVED deprecated config schema classes**
   - Deleted SecretProviderConfig class
   - Deleted PubSubConfig class
   - Deleted InfrastructureConfig class
   - Updated ExecutionContext to use CloudProvider union type

5. **ESTABLISHED single golden path**
   - User calls init() → config_bootstrap.py → ExecutionContext (service-aware clouds) → Session creation
   - No more confusion between old/new patterns
   - All services use consistent service-aware pattern: clouds[].secrets, clouds[].logging, etc.

**RESULT**: The original nb_init() issue is resolved. The dev environment works with clouds[].secrets structure and ExecutionContext now correctly extracts secrets using the service-aware pattern instead of expecting a separate secret_provider field.
